Add trackAssets feature to track all assets in webpack-stats.json file
Asset tracking is off by default, by passing the trackAssets boolean to the
plugin options, all assets output by webpack will be added as a chunk with the
default chunk name 'assets' to the webpack-stats.json file.

This enables rendering of files that are separately processed like gzipped js
and css from 'compress-webpack-plugin'.