Sort properties of `assets` and `chunks` objects.
This change ensures that the `assets` and `chunks` properties of the compilation output are sorted before the JSON is written to disk. This provides an increased degree of determinism in the output file, which is helpful when checking it in or running under a system like Bazel that caches build results. This commit only sorts the `assets` and `chunks` sub-objects because in practice those seem to be the ones most likely to differ.

Fixes issue #69.